Sled: - Improved benchmarking approach - split benchmark & test in dedicated files
